+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+## 2.7.5
+
 ## 2.7.4
 
  - Deactivating a spell bypasses cooldowns and costs.